# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * V75 Analyzer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SegmentRecord {
  id: string;
  timestamp: number;
  date_str: string;
  time_str: string;
  open_price: number;
  close_price: number;
  /** @nullable */
  outcome: string | null;
  point_move: number;
  score: number;
  /** @nullable */
  adjusted_score?: number | null;
  /** @nullable */
  rsi?: number | null;
  /** @nullable */
  stoch_k?: number | null;
  /** @nullable */
  stoch_d?: number | null;
  /** @nullable */
  macd_histogram?: number | null;
  /** @nullable */
  williams_r?: number | null;
  /** @nullable */
  cci?: number | null;
  /** @nullable */
  bb_position?: number | null;
  /** @nullable */
  bb_width?: number | null;
  /** @nullable */
  atr?: number | null;
  /** @nullable */
  relative_atr?: number | null;
  /** @nullable */
  ema9?: number | null;
  /** @nullable */
  ema21?: number | null;
  /** @nullable */
  ema50?: number | null;
  /** @nullable */
  ema200?: number | null;
  /** @nullable */
  ema_alignment?: string | null;
  /** @nullable */
  has_active_bob?: boolean | null;
  /** @nullable */
  has_active_beob?: boolean | null;
  /** @nullable */
  ob_timeframe?: string | null;
  /** @nullable */
  has_fvg_bull?: boolean | null;
  /** @nullable */
  has_fvg_bear?: boolean | null;
  /** @nullable */
  choch_detected?: string | null;
  /** @nullable */
  liquidity_sweep?: string | null;
  /** @nullable */
  h4_bias?: string | null;
  /** @nullable */
  h1_bias?: string | null;
  /** @nullable */
  top_down_alignment?: string | null;
  /** @nullable */
  candle_pattern?: string | null;
  /** @nullable */
  structure?: string | null;
  /** @nullable */
  rsi_divergence?: string | null;
  /** @nullable */
  macd_divergence?: string | null;
  /** @nullable */
  hour_of_day?: number | null;
  /** @nullable */
  day_of_week?: number | null;
  /** @nullable */
  dominant_pattern?: string | null;
  /** @nullable */
  pattern_score?: number | null;
  /** @nullable */
  pattern_direction?: string | null;
  /** @nullable */
  m15_bias?: string | null;
}

export interface SegmentInput {
  timestamp: number;
  date_str: string;
  time_str: string;
  open_price: number;
  close_price: number;
  /** @nullable */
  outcome: string | null;
  point_move: number;
  score: number;
  /** @nullable */
  adjusted_score?: number | null;
  /** @nullable */
  rsi?: number | null;
  /** @nullable */
  stoch_k?: number | null;
  /** @nullable */
  stoch_d?: number | null;
  /** @nullable */
  macd_histogram?: number | null;
  /** @nullable */
  williams_r?: number | null;
  /** @nullable */
  cci?: number | null;
  /** @nullable */
  bb_position?: number | null;
  /** @nullable */
  bb_width?: number | null;
  /** @nullable */
  atr?: number | null;
  /** @nullable */
  relative_atr?: number | null;
  /** @nullable */
  ema9?: number | null;
  /** @nullable */
  ema21?: number | null;
  /** @nullable */
  ema50?: number | null;
  /** @nullable */
  ema200?: number | null;
  /** @nullable */
  ema_alignment?: string | null;
  /** @nullable */
  has_active_bob?: boolean | null;
  /** @nullable */
  has_active_beob?: boolean | null;
  /** @nullable */
  ob_timeframe?: string | null;
  /** @nullable */
  has_fvg_bull?: boolean | null;
  /** @nullable */
  has_fvg_bear?: boolean | null;
  /** @nullable */
  choch_detected?: string | null;
  /** @nullable */
  liquidity_sweep?: string | null;
  /** @nullable */
  h4_bias?: string | null;
  /** @nullable */
  h1_bias?: string | null;
  /** @nullable */
  top_down_alignment?: string | null;
  /** @nullable */
  candle_pattern?: string | null;
  /** @nullable */
  structure?: string | null;
  /** @nullable */
  rsi_divergence?: string | null;
  /** @nullable */
  macd_divergence?: string | null;
  /** @nullable */
  hour_of_day?: number | null;
  /** @nullable */
  day_of_week?: number | null;
  /** @nullable */
  dominant_pattern?: string | null;
  /** @nullable */
  pattern_score?: number | null;
  /** @nullable */
  pattern_direction?: string | null;
  /** @nullable */
  m15_bias?: string | null;
}

export interface SignalRecord {
  id: string;
  timestamp: number;
  direction: string;
  strength: string;
  confidence: number;
  score: number;
  /** @nullable */
  adjusted_score?: number | null;
  duration_minutes: number;
  /** @nullable */
  entry_price?: number | null;
  /** @nullable */
  exit_price?: number | null;
  /** @nullable */
  outcome: string | null;
  /** @nullable */
  choch_present?: boolean | null;
  /** @nullable */
  sweep_present?: boolean | null;
  /** @nullable */
  ob_timeframe?: string | null;
  /** @nullable */
  h4_bias?: string | null;
  /** @nullable */
  h1_bias?: string | null;
  /** @nullable */
  pattern_match_rate?: number | null;
}

export interface SignalInput {
  id: string;
  timestamp: number;
  direction: string;
  strength: string;
  confidence: number;
  score: number;
  /** @nullable */
  adjusted_score?: number | null;
  duration_minutes: number;
  /** @nullable */
  entry_price?: number | null;
  /** @nullable */
  outcome?: string | null;
  /** @nullable */
  choch_present?: boolean | null;
  /** @nullable */
  sweep_present?: boolean | null;
  /** @nullable */
  ob_timeframe?: string | null;
  /** @nullable */
  h4_bias?: string | null;
  /** @nullable */
  h1_bias?: string | null;
  /** @nullable */
  pattern_match_rate?: number | null;
}

export type OutcomeUpdateOutcome = typeof OutcomeUpdateOutcome[keyof typeof OutcomeUpdateOutcome];


export const OutcomeUpdateOutcome = {
  WIN: 'WIN',
  LOSS: 'LOSS',
} as const;

export interface OutcomeUpdate {
  outcome: OutcomeUpdateOutcome;
  exit_price: number;
}

export interface ConditionStat {
  id: string;
  condition_name: string;
  /** @nullable */
  rise_wins?: number | null;
  /** @nullable */
  fall_wins?: number | null;
  /** @nullable */
  total?: number | null;
  /** @nullable */
  accuracy?: number | null;
  weight_multiplier: number;
  /** @nullable */
  updated_at?: string | null;
}

export type ConditionBumpOutcome = typeof ConditionBumpOutcome[keyof typeof ConditionBumpOutcome];


export const ConditionBumpOutcome = {
  RISE: 'RISE',
  FALL: 'FALL',
  FLAT: 'FLAT',
} as const;

export interface ConditionBump {
  conditions: string[];
  outcome: ConditionBumpOutcome;
}

export type ListSegmentsParams = {
limit?: number;
};

export type ListSignalsParams = {
limit?: number;
};