# lib/api-zod/src/generated/types/conditionBump.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * V75 Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { ConditionBumpOutcome } from './conditionBumpOutcome';

export interface ConditionBump {
  conditions: string[];
  outcome: ConditionBumpOutcome;
}