# lib/api-zod/src/generated/types/conditionStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * V75 Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface ConditionStat {
  id: string;
  condition_name: string;
  /** @nullable */
  rise_wins?: number | null;
  /** @nullable */
  fall_wins?: number | null;
  /** @nullable */
  total?: number | null;
  /** @nullable */
  accuracy?: number | null;
  weight_multiplier: number;
  /** @nullable */
  updated_at?: string | null;
}