# lib/api-zod/src/generated/types/outcomeUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * V75 Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { OutcomeUpdateOutcome } from './outcomeUpdateOutcome';

export interface OutcomeUpdate {
  outcome: OutcomeUpdateOutcome;
  exit_price: number;
}